// frontend-seller/src/types/RegisterShopType.ts
export type RegisterShopType = {
    shoeImage: string; // 가게 이미지
    shopName: string; // 가게(상호명)
    shopOwner: string; // 대표자명
    shopAddr: string; // 주소
    shopDetailAddr: string; // 상세주소
    shopIntro: string;  // 가게소개
    shopWorkday: string; // 영업시간
    shopHoliyday: string; // 휴무일
    FoodIngredients: string; // 재료
    FoodOrigin: string; // 원산지
}